[#3725] PR feedback-refactor
